pyproject.toml
[project]
name = "cloud-incident-texts"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pdfplumber>=0.10.0",
    "pillow>=10.0.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "schedule>=1.2.1",
    "trafilatura>=1.8.0",
]

[tool.uv]
dev-dependencies = []
